# perf: parallelize creating commitments, use buffers