--- a/go.mod
+++ b/go.mod
@@ -3,9 +3,10 @@ module github.com/celestiaorg/go-square/v3
 go 1.24.0
 
 require (
-	github.com/celestiaorg/nmt v0.24.2-0.20250918161004-dcaa76b80708
+	github.com/celestiaorg/nmt v0.24.2
 	github.com/stretchr/testify v1.11.1
 	golang.org/x/exp v0.0.0-20231206192017-f3f8817b8deb
+	golang.org/x/sync v0.17.0
 	google.golang.org/protobuf v1.36.9
 )
 
@@ -13,6 +14,5 @@ require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/sync v0.17.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )